move voting power dist cache
