+++ proto/babylon/btcstaking/v1/genesis.proto
@@ -4,7 +4,6 @@ package babylon.btcstaking.v1;
 import "gogoproto/gogo.proto";
 import "babylon/btcstaking/v1/params.proto";
 import "babylon/btcstaking/v1/btcstaking.proto";
-import "babylon/btcstaking/v1/incentive.proto";
 import "babylon/btcstaking/v1/events.proto";
 
 option go_package = "github.com/babylonlabs-io/babylon/x/btcstaking/types";
@@ -23,17 +22,6 @@ message GenesisState {
   repeated BTCDelegator btc_delegators = 6;
   // all the events and its indexes.
   repeated EventIndex events = 7;
-  // vp_dst_cache is the table of all providers voting power with the total at one specific block.
-  // TODO: remove this after not storing in the keeper store it anymore.
-  repeated VotingPowerDistCacheBlkHeight vp_dst_cache = 8;
-}
-
-// VotingPowerDistCacheBlkHeight the total voting power of the finality providers at one specific block height
-message VotingPowerDistCacheBlkHeight {
-  // block_height is the height of the block the voting power distribution cached was stored.
-  uint64 block_height = 1;
-  // vp_distribution the finality providers distribution cache at that height.
-  VotingPowerDistCache vp_distribution = 2;
 }
 
 // BlockHeightBbnToBtc stores the btc <-> bbn block.
